Draw strokes on pointer moves.
When the pointer moves, we used to see a cicle at the beginning and the end of
the movement. Now, we draw a convex shape which represents the stroke, where the
stroke width varies with the pressure of the events.

This is not yet a smooth drawing, as we only differentiate onces, but this
already gives some continuity to the pictures.