License this project as MPL v2.0.
This project is aimed to remain open for anyone to use and contribute, for the
benefit of a greater good.

On the other hand, making free software should not be a limitation for anybody
who might seize opportunity to provide better services for end users. As such,
this license should allow embedders who might want to make earn money in order
to provide and maintain a service based on this software.